fix: harden operational alert outbox delivery

+++ src/main/java/com/example/attend/operations/infrastructure/mybatis/FinalizationOperationalEventMapper.java
@@ -10,7 +10,7 @@ import org.apache.ibatis.annotations.Param;
 @Mapper
 public interface FinalizationOperationalEventMapper {
 
-    int insertRetryExhaustedEvent(
+    Long insertRetryExhaustedEvent(
             @Param("attendanceDayId") long attendanceDayId,
             @Param("incidentClaimVersion") long incidentClaimVersion,
             @Param("errorCode") String errorCode,
